use and move specialfield lock component
